Fix issue of skipping characters when pasting in text.
I believe this is happening when pasting since keys are entering the buffer
faster than were' pulling them out, so there's a new one available when
calling _read_kbd(0,1,0) and that one is returned instead of repeating the
last discarded one that was saved in waiting_char.

This fixes bash where you can paste into the OS/2 terminal, also tested with
rlbasic.exe.